.github/workflows/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
xhtml2pdf
google-genai
python-dotenv
